sycl_context deprecation workaround (#1423)
* sycl_context deprecation workaround

* fix exception logic, sklearn compatibility

* moving daal4py/sycl examples to tests

* adding from_python arg, keeping sycl_context

* remove duplicate stream

* add read_csv to streaming as well

* removing default arg from context

* file formatting

* address codefactor except

* additional address codefactor

* add version check

* change ONEDAL_VERSION to INTEL_DAAL_VERSION
