+++ conda-recipe/meta.yaml
@@ -81,7 +81,7 @@ test:
         - python -m unittest discover -v -s tests -p test*.py
         - pytest --pyargs daal4py/sklearn/
         - python tests/run_examples.py
-        - python -m daal4py examples/daal4py/sycl/sklearn_sycl.py
+        - python -m daal4py tests/daal4py/sycl/sklearn_sycl.py
 
 about:
     about:
