+++ src/oneapi/oneapi_backend.cpp
@@ -23,11 +23,15 @@ static_assert(false, "DAAL_SYCL_INTERFACE not defined")
 
 #include "oneapi_backend.h"
 
-    PySyclExecutionContext::PySyclExecutionContext(const std::string & dev)
+    PySyclExecutionContext::PySyclExecutionContext(const std::string & dev, const bool from_python)
     : m_ctxt(NULL)
 {
     if (dev == "gpu")
+#if INTEL_DAAL_VERSION >= 20240000
+        m_ctxt = new daal::services::SyclExecutionContext(cl::sycl::queue(cl::sycl::gpu_selector()), from_python);
+#else // INTEL_DAAL_VERSION >= 20240000
         m_ctxt = new daal::services::SyclExecutionContext(cl::sycl::queue(cl::sycl::gpu_selector()));
+#endif // INTEL_DAAL_VERSION >= 20240000
     else if (dev == "cpu")
         m_ctxt = new daal::services::SyclExecutionContext(cl::sycl::queue(cl::sycl::cpu_selector()));
     else if (dev == "host")
