+++ tests/daal4py/sycl/covariance.py
@@ -41,7 +41,7 @@ try:
 
     with sycl_context("gpu"):
         gpu_available = True
-except:
+except Exception:
     gpu_available = False
 
 
@@ -72,7 +72,16 @@ def to_numpy(data):
 
 
 def main(readcsv=read_csv, method="defaultDense"):
-    infile = os.path.join("..", "data", "batch", "covcormoments_dense.csv")
+    infile = os.path.join(
+        "..",
+        "..",
+        "..",
+        "examples",
+        "daal4py",
+        "data",
+        "batch",
+        "covcormoments_dense.csv",
+    )
 
     # Load the data
     data = readcsv(infile, range(10), t=np.float32)
