+++ tests/daal4py/sycl/covariance_streaming.py
@@ -27,15 +27,50 @@ import daal4py as d4p
 from daal4py.oneapi import sycl_buffer
 
 sys.path.insert(0, "..")
-from stream import read_next
 
 try:
     from daal4py.oneapi import sycl_context
 
     with sycl_context("gpu"):
         gpu_available = True
-except:
+except Exception:
     gpu_available = False
+
+try:
+    import pandas
+
+    def read_csv(f, c=None, s=0, n=None, t=np.float64):
+        return pandas.read_csv(
+            f, usecols=c, delimiter=",", header=None, skiprows=s, nrows=n, dtype=t
+        )
+
+except Exception:
+    # fall back to numpy genfromtxt
+    def read_csv(f, c=None, s=0, n=np.iinfo(np.int64).max):
+        a = np.genfromtxt(f, usecols=c, delimiter=",", skip_header=s, max_rows=n)
+        if a.shape[0] == 0:
+            raise Exception("done")
+        if a.ndim == 1:
+            return a[:, np.newaxis]
+        return a
+
+
+# a generator which reads a file in chunks
+def read_next(file, chunksize, readcsv=read_csv):
+    assert os.path.isfile(file)
+    s = 0
+    while True:
+        # if found a smaller chunk we set s to < 0 to indicate eof
+        if s < 0:
+            return
+        a = read_csv(file, s=s, n=chunksize)
+        # last chunk is usually smaller, if not,
+        # numpy will print warning in next iteration
+        if chunksize > a.shape[0]:
+            s = -1
+        else:
+            s += a.shape[0]
+        yield a
 
 
 # At this moment with sycl we are working only with numpy arrays
@@ -58,7 +93,16 @@ def to_numpy(data):
 
 
 def main(readcsv=None, method="defaultDense"):
-    infile = os.path.join("..", "data", "batch", "covcormoments_dense.csv")
+    infile = os.path.join(
+        "..",
+        "..",
+        "..",
+        "examples",
+        "daal4py",
+        "data",
+        "batch",
+        "covcormoments_dense.csv",
+    )
 
     # Using of the classic way (computations on CPU)
     # configure a covariance object
