+++ tests/daal4py/sycl/decision_forest_classification.py
@@ -14,7 +14,7 @@
 # limitations under the License.
 # ==============================================================================
 
-# daal4py Decision Forest Regression example for shared memory systems
+# daal4py Decision Forest Classification example for shared memory systems
 
 import os
 
@@ -47,14 +47,17 @@ except Exception:
 
 # Commone code for both CPU and GPU computations
 def compute(train_data, train_labels, predict_data, method="defaultDense"):
-    # Configure a training object
-    train_algo = d4p.decision_forest_regression_training(
-        nTrees=100,
+    # Configure a training object (5 classes)
+    train_algo = d4p.decision_forest_classification_training(
+        5,
         fptype="float",
-        engine=d4p.engines_mt2203(seed=777),
-        varImportance="MDA_Raw",
+        nTrees=10,
+        minObservationsInLeafNode=8,
+        featuresPerNode=3,
+        engine=d4p.engines_mt19937(seed=777),
+        varImportance="MDI",
         bootstrap=True,
-        resultsToCompute="computeOutOfBagError|computeOutOfBagErrorPerObservation",
+        resultsToCompute="computeOutOfBagError",
         method=method,
     )
     # Training result provides (depending on parameters) model,
@@ -62,7 +65,12 @@ def compute(train_data, train_labels, predict_data, method="defaultDense"):
     train_result = train_algo.compute(train_data, train_labels)
 
     # now predict using the model from the training above
-    predict_algo = d4p.decision_forest_regression_prediction(fptype="float")
+    predict_algo = d4p.decision_forest_classification_prediction(
+        nClasses=5,
+        fptype="float",
+        resultsToEvaluate="computeClassLabels|computeClassProbabilities",
+        votingMethod="unweighted",
+    )
 
     predict_result = predict_algo.compute(predict_data, train_result.model)
 
@@ -89,10 +97,28 @@ def to_numpy(data):
 
 
 def main(readcsv=read_csv, method="defaultDense"):
-    nFeatures = 13
+    nFeatures = 3
     # input data file
-    train_file = os.path.join("..", "data", "batch", "df_regression_train.csv")
-    predict_file = os.path.join("..", "data", "batch", "df_regression_test.csv")
+    train_file = os.path.join(
+        "..",
+        "..",
+        "..",
+        "examples",
+        "daal4py",
+        "data",
+        "batch",
+        "df_classification_train.csv",
+    )
+    predict_file = os.path.join(
+        "..",
+        "..",
+        "..",
+        "examples",
+        "daal4py",
+        "data",
+        "batch",
+        "df_classification_test.csv",
+    )
 
     # Read train data. Let's use 3 features per observation
     train_data = readcsv(train_file, range(nFeatures), t=np.float32)
@@ -106,7 +132,7 @@ def main(readcsv=read_csv, method="defaultDense"):
         train_data, train_labels, predict_data, "defaultDense"
     )
     assert predict_result.prediction.shape == (predict_labels.shape[0], 1)
-    assert (np.square(predict_result.prediction - predict_labels).mean() < 18).any()
+    assert (np.mean(predict_result.prediction != predict_labels) < 0.03).any()
 
     train_data = to_numpy(train_data)
     train_labels = to_numpy(train_labels)
@@ -122,9 +148,7 @@ def main(readcsv=read_csv, method="defaultDense"):
                 sycl_train_data, sycl_train_labels, sycl_predict_data, "hist"
             )
             assert predict_result.prediction.shape == (predict_labels.shape[0], 1)
-            assert (
-                np.square(predict_result.prediction - predict_labels).mean() < 18
-            ).any()
+            assert (np.mean(predict_result.prediction != predict_labels) < 0.03).any()
 
     return (train_result, predict_result, predict_labels)
 
@@ -136,6 +160,10 @@ if __name__ == "__main__":
     print(
         "\nDecision forest prediction results (first 10 rows):\n",
         predict_result.prediction[0:10],
+    )
+    print(
+        "\nDecision forest probabilities results (first 10 rows):\n",
+        predict_result.probabilities[0:10],
     )
     print("\nGround truth (first 10 rows):\n", plabels[0:10])
     print("All looks good!")
