+++ tests/daal4py/sycl/kmeans.py
@@ -41,7 +41,7 @@ try:
 
     with sycl_context("gpu"):
         gpu_available = True
-except:
+except Exception:
     gpu_available = False
 
 
@@ -83,7 +83,9 @@ def to_numpy(data):
 
 
 def main(readcsv=read_csv, method="randomDense"):
-    infile = os.path.join("..", "data", "batch", "kmeans_dense.csv")
+    infile = os.path.join(
+        "..", "..", "..", "examples", "daal4py", "data", "batch", "kmeans_dense.csv"
+    )
     nClusters = 20
     maxIter = 5
