+++ tests/daal4py/sycl/log_reg_binary_dense.py
@@ -14,7 +14,7 @@
 # limitations under the License.
 # ==============================================================================
 
-# daal4py BF KNN example for shared memory systems
+# daal4py logistic regression example for shared memory systems
 
 import os
 
@@ -41,8 +41,20 @@ try:
 
     with sycl_context("gpu"):
         gpu_available = True
-except:
+except Exception:
     gpu_available = False
+
+
+# Commone code for both CPU and GPU computations
+def compute(train_data, train_labels, predict_data, nClasses):
+    # set parameters and train
+    train_alg = d4p.logistic_regression_training(
+        nClasses=nClasses, interceptFlag=True, fptype="float"
+    )
+    train_result = train_alg.compute(train_data, train_labels)
+    # set parameters and compute predictions
+    predict_alg = d4p.logistic_regression_prediction(nClasses=nClasses, fptype="float")
+    return predict_alg.compute(predict_data, train_result.model), train_result
 
 
 # At this moment with sycl we are working only with numpy arrays
@@ -64,72 +76,70 @@ def to_numpy(data):
     return data
 
 
-# Common code for both CPU and GPU computations
-def compute(train_data, train_labels, predict_data, nClasses):
-    # Create an algorithm object and call compute
-    train_algo = d4p.bf_knn_classification_training(nClasses=nClasses, fptype="float")
-    train_result = train_algo.compute(train_data, train_labels)
-
-    # Create an algorithm object and call compute
-    predict_algo = d4p.bf_knn_classification_prediction(nClasses=nClasses, fptype="float")
-    predict_result = predict_algo.compute(predict_data, train_result.model)
-    return predict_result
-
-
 def main(readcsv=read_csv, method="defaultDense"):
-    # Input data set parameters
-    train_file = os.path.join("..", "data", "batch", "k_nearest_neighbors_train.csv")
-    predict_file = os.path.join("..", "data", "batch", "k_nearest_neighbors_test.csv")
+    nClasses = 2
+    nFeatures = 20
 
-    # Read data. Let's use 5 features per observation
-    nFeatures = 5
-    nClasses = 5
-    train_data = readcsv(train_file, range(nFeatures), t=np.float32)
-    train_labels = readcsv(train_file, range(nFeatures, nFeatures + 1), t=np.float32)
-    predict_data = readcsv(predict_file, range(nFeatures), t=np.float32)
-    predict_labels = readcsv(predict_file, range(nFeatures, nFeatures + 1), t=np.float32)
+    # read training data from file with 20 features per observation and 1 class label
+    trainfile = os.path.join(
+        "..", "..", "..", "examples", "daal4py", "data", "batch", "binary_cls_train.csv"
+    )
+    train_data = readcsv(trainfile, range(nFeatures), t=np.float32)
+    train_labels = readcsv(trainfile, range(nFeatures, nFeatures + 1), t=np.float32)
 
-    predict_result_classic = compute(train_data, train_labels, predict_data, nClasses)
+    # read testing data from file with 20 features per observation
+    testfile = os.path.join(
+        "..", "..", "..", "examples", "daal4py", "data", "batch", "binary_cls_test.csv"
+    )
+    predict_data = readcsv(testfile, range(nFeatures), t=np.float32)
+    predict_labels = readcsv(testfile, range(nFeatures, nFeatures + 1), t=np.float32)
 
-    # We expect less than 170 mispredicted values
-    assert np.count_nonzero(predict_labels != predict_result_classic.prediction) < 170
+    # Using of the classic way (computations on CPU)
+    result_classic, train_result = compute(
+        train_data, train_labels, predict_data, nClasses
+    )
 
     train_data = to_numpy(train_data)
     train_labels = to_numpy(train_labels)
     predict_data = to_numpy(predict_data)
 
+    # It is possible to specify to make the computations on GPU
     if gpu_available:
         with sycl_context("gpu"):
             sycl_train_data = sycl_buffer(train_data)
             sycl_train_labels = sycl_buffer(train_labels)
             sycl_predict_data = sycl_buffer(predict_data)
-
-            predict_result_gpu = compute(
+            result_gpu, _ = compute(
                 sycl_train_data, sycl_train_labels, sycl_predict_data, nClasses
             )
-            assert np.allclose(
-                predict_result_gpu.prediction, predict_result_classic.prediction
-            )
 
+        assert np.mean(result_classic.prediction != result_gpu.prediction) < 0.2
+
+    # It is possible to specify to make the computations on GPU
     with sycl_context("cpu"):
         sycl_train_data = sycl_buffer(train_data)
         sycl_train_labels = sycl_buffer(train_labels)
         sycl_predict_data = sycl_buffer(predict_data)
-
-        predict_result_cpu = compute(
+        result_cpu, _ = compute(
             sycl_train_data, sycl_train_labels, sycl_predict_data, nClasses
         )
-        assert np.allclose(
-            predict_result_cpu.prediction, predict_result_classic.prediction
-        )
 
-    return (predict_result_classic, predict_labels)
+    # the prediction result provides prediction
+    assert result_classic.prediction.shape == (
+        predict_data.shape[0],
+        train_labels.shape[1],
+    )
+
+    assert np.mean(result_classic.prediction != result_cpu.prediction) < 0.2
+    return (train_result, result_classic, predict_labels)
 
 
 if __name__ == "__main__":
-    (predict_result, predict_labels) = main()
-    print("BF based KNN classification results:")
-    print("Ground truth(observations #30-34):\n", predict_labels[30:35])
+    (train_result, predict_result, predict_labels) = main()
+    print("\nLogistic Regression coefficients:\n", train_result.model.Beta)
     print(
-        "Classification results(observations #30-34):\n", predict_result.prediction[30:35]
+        "\nLogistic regression prediction results (first 10 rows):\n",
+        predict_result.prediction[0:10],
     )
+    print("\nGround truth (first 10 rows):\n", predict_labels[0:10])
+    print("All looks good!")
