+++ tests/daal4py/sycl/log_reg_dense.py
@@ -41,7 +41,7 @@ try:
 
     with sycl_context("gpu"):
         gpu_available = True
-except:
+except Exception:
     gpu_available = False
 
 
@@ -90,12 +90,16 @@ def main(readcsv=read_csv, method="defaultDense"):
     nFeatures = 6
 
     # read training data from file with 6 features per observation and 1 class label
-    trainfile = os.path.join("..", "data", "batch", "logreg_train.csv")
+    trainfile = os.path.join(
+        "..", "..", "..", "examples", "daal4py", "data", "batch", "logreg_train.csv"
+    )
     train_data = readcsv(trainfile, range(nFeatures), t=np.float32)
     train_labels = readcsv(trainfile, range(nFeatures, nFeatures + 1), t=np.float32)
 
     # read testing data from file with 6 features per observation
-    testfile = os.path.join("..", "data", "batch", "logreg_test.csv")
+    testfile = os.path.join(
+        "..", "..", "..", "examples", "daal4py", "data", "batch", "logreg_test.csv"
+    )
     predict_data = readcsv(testfile, range(nFeatures), t=np.float32)
 
     # Using of the classic way (computations on CPU)
